QTrator/requirements.txt
numpy
qiskit
PyYAML
pandas
matplotlib
z3-solver
//optional
//qiskit-aer-gpu
pylatexenc